Fixing crashing bug with new thread for manual path recording play. Helps #3
